Neural Net as class/ function fit demo
Set up demo for neural networks to fit to arbitrary functions, able to easily compare hyperparameters, nn shapes, and more visually by watching how they do as the epochs go by.

This was done initially with individual inputs in for loops, then followed by making it work in numpy matrix operations. Applying this to the mnist fashion data was the assignment.

After this, I organized the neural network into a (somewhat disorganized) class, and created a small demo for function fitting with it.